extract runnable scripts from main parsing engine
