+++ lib/decoders.js
