+++ package.json
@@ -2,7 +2,7 @@
   "name": "heroprotocoljs",
   "version": "0.2.1",
   "description": "Javascript port of the heroprotocol Python library to decode Heroes of the Storm replay protocols.",
-  "main": "heroprotocol.js",
+  "main": "index.js",
   "dependencies": {
     "mpyqjs": "^1.0.1",
     "yargs": "^3.32.0"
